remove internal building of lvm2.
genkernel now uses lvm2.static from the base system. Funtoo Linux always
builds lvm.static so there is no need for genkernel to do this anymore. Also,
the "make install" target was trying to create /etc/lvm on the live filesystem
which caused a sandbox violation when genkernel is used inside an ebuild
(and is just a bad thing to happen in general.)