Proper AUFS support
This patch is what we have been using for the Pentoo LiveCDs for about 5
years with the changes suggested by lxnay merged in.  As much as
possible has been addressed (with the exception of the man page
documentation).  Please review this code as the next version will likely
be suggested for inclusion (once I write the documentation for the new
features).

Signed-off-by: Rick Farina (Zero_Chaos) <zerochaos@gentoo.org>