Make --kerncache compatible with recent versions of GNU tar, bug #457168
Older versions of GNU tar would extract files to the interior of the
directory specified after an adjustment via -C was taken into
consideration. Newer versions will perform extraction to the exterior of
that directory. This caused --kerncache to override the /lib symlink
with an actual directory.

We correct this by switching to a more standard syntax that should be
consistent against all tar implementations, including both old and new
versions of GNU tar.

Signed-off-by: Richard Yao <ryao@gentoo.org>