Enable CONFIG_AUDIT=y on almost all architectures
CONFIG_AUDIT is already enabled on the most common architectures. It is
ancient, yet it is required by SELinux. Lets enable it on all
architectures for SELinux compatibility.

It is not clear that ./arch/ppc64/kernel-2.6-pSeries and
./arch/ppc64/kernel-2.6.g5 even still work. I am omitting them from this
commit because it is not clear to me that they even still work.  Someone
with appropriate hardware will need to update the config file(s).

Signed-off-by: Richard Yao <ryao@gentoo.org>