Initial kernel config infrastructure
The merge.pl script and initial configuration files for ppc, ppc64 x86
and x86_64 are from RHEL6. The merge.pl is intended to be replaced in
the future. We will use it for now because Rick Farina of Gentoo QA
wants this merged regardless of whether we have a substitute ready or
not.

Signed-off-by: Richard Yao <ryao@gentoo.org>