Disable stack protection by default
I had intended for this feature to catch bugs in the kernel that
involved stack corruption so that there would be a chance of debugging
them, but Rick Farina is strongly against turning this on unless it is
the strong version, which we cannot turn on until GCC 4.9+ becomes
widespread. I am turning this off because the controversy is simply not
worth it, but this is a bad decision.

Signed-off-by: Richard Yao <ryao@gentoo.org>