Fix build issues with the dmraid module
Attempts at building an initramfs with --dmraid have started to fail very
recently. The failure itself seems to be due to changes to the current
system versions of lvm2, that have brought to light the fact that the
genkernel dmraid has been building using /usr/include/ headers instead of
headers from the genkernel lvm2.  Additionally to this, dmraid requires
libdevmapper-event, and the genkernel version of lvm2 doesn't (didn't ever?)
build this.

This commit addresses both issues, by:

1. Adding --enable-dmeventd --enable-cmdlib to the lvm configuration, so
that libdevmapper-event is available to dmraid

2. Adding a --with-devmappert-prefix= to dmraid configure so that
./configure won't try and override it or otherwise get in the way of
the hard-coded paths we add to CFLAGS.

3. Adding a DEVMAPPEREVENT_CFLAGS var as dmraid ./configure erroneously
does not set this var to anything, despite it being used later.

Received nod from Zero_Chaos for the commit.