fix rootfs detection per bug #552052
ZFS and btrfs tools if not explictly configured (either way), default to
detecting the fstype of the rootfs.

The old rootfs detection method was NOT reliable in all cases; so
replace it with a different way. Future improvements should check for
rootfs AND the local existence of the required binaries (and bail out
early if the the rootrfs was detected but the binaries are not present).

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>