always respect kernel settings
Just because someone has passed the cross-compile flag does not mean
the other settings are not interesting as well.  The cross-compile
prefix is used for many tools, but genkernel only allows for setting
of a few, and the other settings (like arch and output dir) should
not be completely ignored.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>