adjust System.map die check for #570822
Builds with newer genkernel have been failing in catalyst like so:
* ERROR: Could not copy System.map from kerncache for the kernel package!

The code that added it seems to have been meant to check the minkernpackage/
subdir where the file was renamed instead of the current subdir.  Change the
check to simplify verify the mv didn't fail instead of duplicating the file
name entirely.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>