busybox: add more functionality, pass 1.
While trying to install a difficult system, it became apparent that the
default genkernel busybox was a pain to work with in some cases, and the
netboot busybox hadn't been maintained. The netboot busybox
functionality also adds very little size to busybox, so start to merge
the busybox configurations (meaning default genkernel can be used for
most of netboot).

On amd64, this adds approx 170KiB to the default busybox, and 380KiB to
the netboot busybox (mostly due to not being maintained much).

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>