Include libgcc_s.so.1 with ZFS support
The resumable send/recv work added an indirect dependency on
libgcc_s.so.1, which libpthread opens via dlopen for _Unwind_Resume().
If we do not include it, we will see:

libgcc_s.so.1 must be installed for pthread_cancel to work

Lets include it for now. We make an effort to include only 1 version.
This might not work on all architectures, but for lack of ability to
test, it will need to do for now. It also will use gcc-config if it is
avaliable.

This also enables functions in zdb that previously did not work due to
the same issue.

Original-patch-by: Jason Zaman <perfinion@gentoo.org>
Signed-off-by: Richard Yao <ryao@gentoo.org>