linuxrc: fix init test for cdroot/squashfs.
In a cdroot-squashfs environment, the /sbin/init path is an absolute
symlink that is ONLY valid inside the chroot. As such, doing a
executable test -x on it, from outside the chroot will fail, because the
symlink destination is not valid outside $CHROOT.

Improve the test to be done inside the chroot. Alternatively, it could
detect a symlink and dereference it inside the chroot.

Closes: https://bugs.gentoo.org/639034
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>