Preliminary support for systemd-boot bootloader.
This adds preliminary support for systemd-boot using the kernel-install
tool.  It's dependent upon systemd pull request #11281
(https://github.com/systemd/systemd/pull/11281) to work properly with
initrd files.

It's a little ugly because it wipes (to avoid double copying files) it
wipes out the ones that genkernel has already copied to /boot, since
genkernel copies those in multiple different places in the code.

Ideally those would be refactored and delayed until everything had built
successfully, but that was a large refactoring, so for now this adds
support as is.

Signed-off-by: Mike Auty <mike.auty@gmail.com>