ZFS: Change hostid handling
No need to deal with hostid value: We assume that /etc/hostid exists if ZFS
is used. So just copy that existing file.

If file is missing tell user that kernel command-line parameter "spl_hostid=<hostid>"
must be used. Same is true when generated initramfs is used for multiple systems.

Closes: https://bugs.gentoo.org/531192
Closes: https://bugs.gentoo.org/616738
Signed-off-by: Thomas Deutschmann <whissi@gentoo.org>