Add genkernel worker module "gkbuild"
In the future, genkernel will try to compile all programs used in
genkernel's initramfs on its own instead of copying from host system.

This commit will add support for gkbuilds.

A "gkbuild" is like an ebuild but with reduced functionality.

Using an ebuild-like system will make it easier to maintain those programs
in a standardized way, especially with reworked cross-compile support.

Signed-off-by: Thomas Deutschmann <whissi@gentoo.org>