linuxrc: Change ROOTDELAY handling
Before this change we converted $ROOTDELAY into microseconds and
decremented that value on each loop iteration. Once that value was
<=0 we threw a timeout.

Because we substracted a fixed value we didn't take into account that
a command from loop could already have taken some time. During testing,
in worst case, running with ROOTDELAY=10 and invalid root=UUID= parameter,
it was seen that it took up to ~35s instead of 10s before linuxrc
prompted for new root value.

With this change, we now set a timeout based on current time in seconds
+ ROOTDELAY. Loop will end if current time is >= timeout.

In addition, ROOTDELAY default value was changed from 1 to 5.

Signed-off-by: Thomas Deutschmann <whissi@gentoo.org>