linuxrc: Load network modules only when needed
To avoid problems related to drivers requiring special firmware which
might be not available when loading the module because the user don't really
need that module but it was added based on genkernel's module_load file,
we will no longer load network modules on boot.

Instead we will only load network modules when needed, for example
when dosshd is set or NFS is used.

Signed-off-by: Thomas Deutschmann <whissi@gentoo.org>