initrd.scripts: prompt_user(): Read answer from $GK_PROMPT_FILE on timeout
This commit will bring back read timeout which was removed in
Commit a280829. Read timeout is still disabled by default but
can now be controlled via gk.prompt.timeout kernel command-line
argument.

When gk.prompt.timeout is != 0, the current prompt will be written
to $GK_PROMPT_FILE. This will allow remote user to notice that
there's a prompt.

In addition, on timeout, $GK_PROMPT_FILE will be read allowing
remote user to answer prompt and hopefully to fix the problem
and resume booting.

To make this work, gk.prompt.timeout will be set to 30 seconds
by default when SSH usage is enabled.

Signed-off-by: Thomas Deutschmann <whissi@gentoo.org>