Don't compile multipath-tools; Copy from host
multipath-tools package doesn't really support a static
build: While we managed to build programs like multipath
statically, it still relied on dynamic libs for checkers
which we are unable to fix.

This commit will change --multipath support in that way that
we now require that sys-fs/multipath-tools is installed on
host system because we will copy binaries and used libraries
from host to initramfs.

Due to this change, you can't add multipath support in cross-
compile scenarios anymore.

Bug: https://bugs.gentoo.org/695510
Signed-off-by: Thomas Deutschmann <whissi@gentoo.org>