defaults/linuxrc: Process /etc/initramfs.mounts multiple times
We need to mount mountpoints listed in /etc/initramfs.mounts
before validating REAL_INIT in case init is located on seperate
mount.

In addition the code was moved to a dedicated function named
process_initramfs_mounts() to allow to run it multiple times
which is needed if REAL_INIT wasn't verified (in case system
was booted from livecd).

Link: https://forums.gentoo.org/viewtopic-t-1117762.html
Signed-off-by: Thomas Deutschmann <whissi@gentoo.org>