gen_configkernel.sh: config_kernel(): Warn when kernel config did not change
When user requests to modify kernel config, i.e. `genkernel --menuconfig`
but kernel config did not change, show a warning to allow user to notice.

This can happen for example when user loaded a different kernel config
in dialog but forgot to save to $KERNEL_OUTPUTDIR/.config which is
required in that case.

Link 1: https://forums.gentoo.org/viewtopic-p-8588232.html#8588232
Signed-off-by: Thomas Deutschmann <whissi@gentoo.org>