gen_compile.sh: populate_binpkg(): Throw away binpkg when glibc has changed
It was reported that dropbear in genkernel initramfs crashed when
client tried to connect.

The problem is, that while we are compiling statically, we still depend
on used glibc version. I.e. dropbear binpkg was created against glibc
version A, but when genkernel initramfs will be created, we will copy
system libraries from whatever glibc version host is currently using which
doesn't necessarily need to match used glibc version when binpkg was
created which could result in reported problem.

This commit will make genkernel treat binpkg as stale when glibc was
changed since binpkg was created to force a rebuild.

Signed-off-by: Thomas Deutschmann <whissi@gentoo.org>