gen_compile.sh: populate_binpkg(): Use libc.a for testing if glibc has changed
NSS files module was merged into libc in >=glibc-2.34 so it can't be used
to test for changed glibc version anymore.

Using libc.a seems to be the better alternative which we use already in
get_chost_libdir() and depend on in general.

In addition, a check was added to ensure that our glibc test file really
exists (older-than check would pass if our test file would be missing
which is not what we want).

Fixes: 61049d29 ("gen_compile.sh: populate_binpkg(): Throw away binpkg when glibc has changed")
Bug: https://bugs.gentoo.org/823716
Signed-off-by: Thomas Deutschmann <whissi@gentoo.org>